Provide additional protection against key name leaks
The existence of a key name could be discovered by a timing attack, giving a
very large input and seeing if the "invalid hash" comes back quickly or not.
Mitigate this by, upon invalid key name, performing a dummy HMAC SHA256
against the input before returning the error.

Also, provide extra options in the form of $wgSecureHTMLSpecialDropdown
(default True) for Special:SecureHTML.  If set to False, the key name input
field becomes a text field instead of a dropdown.

Thanks to Alyssa Milburn for discovering the potential timing attack.

Change-Id: Id46419aa4a3a5dd6bc7fe9f8d5df7555d29c169f